Simplified call signature of get_required, fixed PEP8 issues and renamed _derive directory
