+++ esmvaltool/preprocessor/_derive/cllmtisccp.py
@@ -1,14 +1,14 @@
-"""Derivation of variable `clmtkisccp`."""
+"""Derivation of variable `cllmtisccp`."""
 
 
 import iris
 from iris import Constraint
 
-from ._derived_variable import DerivedVariable
+from ._derived_variable_base import DerivedVariableBase
 
 
-class clmtkisccp(DerivedVariable):  # noqa
-    """Derivation of variable `clmtkisccp`."""
+class DerivedVariable(DerivedVariableBase):
+    """Derivation of variable `cllmtisccp`."""
 
     def get_required(self, frequency):
         """Get variable `short_name` and `field` pairs required for derivation.
@@ -28,7 +28,7 @@ class clmtkisccp(DerivedVariable):  # noqa
         return [('clisccp', 'T4' + frequency)]
 
     def calculate(self, cubes):
-        """Compute ISCCP middle level thick cloud area fraction.
+        """Compute ISCCP low level medium-thickness cloud area fraction.
 
         Parameters
         ----------
@@ -38,26 +38,27 @@ class clmtkisccp(DerivedVariable):  # noqa
         Returns
         -------
         iris.cube.Cube
-            `Cube` containing ISCCP middle level thick cloud area fraction.
+            `Cube` containing ISCCP low level medium-thickness cloud area
+            fraction.
 
         """
         clisccp_cube = cubes.extract_strict(
             Constraint(name='isccp_cloud_area_fraction'))
 
         tau = iris.Constraint(
-            atmosphere_optical_thickness_due_to_cloud=lambda t: t > 23.)
-        plev = iris.Constraint(air_pressure=lambda p: 44000. < p <= 68000.)
-        clmtkisccp_cube = clisccp_cube
-        clmtkisccp_cube = clmtkisccp_cube.extract(tau & plev)
+            atmosphere_optical_thickness_due_to_cloud=lambda t: 3.6 < t <= 23.)
+        plev = iris.Constraint(air_pressure=lambda p: p > 68000.)
+        cllmtisccp_cube = clisccp_cube
+        cllmtisccp_cube = cllmtisccp_cube.extract(tau & plev)
         coord_names = [
-            coord.standard_name for coord in clmtkisccp_cube.coords()
+            coord.standard_name for coord in cllmtisccp_cube.coords()
             if len(coord.points) > 1
         ]
         if 'atmosphere_optical_thickness_due_to_cloud' in coord_names:
-            clmtkisccp_cube = clmtkisccp_cube.collapsed(
+            cllmtisccp_cube = cllmtisccp_cube.collapsed(
                 'atmosphere_optical_thickness_due_to_cloud', iris.analysis.SUM)
         if 'air_pressure' in coord_names:
-            clmtkisccp_cube = clmtkisccp_cube.collapsed('air_pressure',
+            cllmtisccp_cube = cllmtisccp_cube.collapsed('air_pressure',
                                                         iris.analysis.SUM)
 
-        return clmtkisccp_cube
+        return cllmtisccp_cube
