+++ esmvaltool/preprocessor/_derive/clmtkisccp.py
@@ -1,14 +1,14 @@
-"""Derivation of variable `clhtkisccp`."""
+"""Derivation of variable `clmtkisccp`."""
 
 
 import iris
 from iris import Constraint
 
-from ._derived_variable import DerivedVariable
+from ._derived_variable_base import DerivedVariableBase
 
 
-class clhtkisccp(DerivedVariable):  # noqa
-    """Derivation of variable `clhtkisccp`."""
+class DerivedVariable(DerivedVariableBase):
+    """Derivation of variable `clmtkisccp`."""
 
     def get_required(self, frequency):
         """Get variable `short_name` and `field` pairs required for derivation.
@@ -28,7 +28,7 @@ class clhtkisccp(DerivedVariable):  # noqa
         return [('clisccp', 'T4' + frequency)]
 
     def calculate(self, cubes):
-        """Compute ISCCP high level thick cloud area fraction.
+        """Compute ISCCP middle level thick cloud area fraction.
 
         Parameters
         ----------
@@ -38,7 +38,7 @@ class clhtkisccp(DerivedVariable):  # noqa
         Returns
         -------
         iris.cube.Cube
-            `Cube` containing ISCCP high level thick cloud area fraction.
+            `Cube` containing ISCCP middle level thick cloud area fraction.
 
         """
         clisccp_cube = cubes.extract_strict(
@@ -46,18 +46,18 @@ class clhtkisccp(DerivedVariable):  # noqa
 
         tau = iris.Constraint(
             atmosphere_optical_thickness_due_to_cloud=lambda t: t > 23.)
-        plev = iris.Constraint(air_pressure=lambda p: p <= 44000.)
-        clhtkisccp_cube = clisccp_cube
-        clhtkisccp_cube = clhtkisccp_cube.extract(tau & plev)
+        plev = iris.Constraint(air_pressure=lambda p: 44000. < p <= 68000.)
+        clmtkisccp_cube = clisccp_cube
+        clmtkisccp_cube = clmtkisccp_cube.extract(tau & plev)
         coord_names = [
-            coord.standard_name for coord in clhtkisccp_cube.coords()
+            coord.standard_name for coord in clmtkisccp_cube.coords()
             if len(coord.points) > 1
         ]
         if 'atmosphere_optical_thickness_due_to_cloud' in coord_names:
-            clhtkisccp_cube = clhtkisccp_cube.collapsed(
+            clmtkisccp_cube = clmtkisccp_cube.collapsed(
                 'atmosphere_optical_thickness_due_to_cloud', iris.analysis.SUM)
         if 'air_pressure' in coord_names:
-            clhtkisccp_cube = clhtkisccp_cube.collapsed('air_pressure',
+            clmtkisccp_cube = clmtkisccp_cube.collapsed('air_pressure',
                                                         iris.analysis.SUM)
 
-        return clhtkisccp_cube
+        return clmtkisccp_cube
