--- esmvaltool/preprocessor/derived_variables/__init__.py
@@ -1,1 +1,0 @@
-"""Automatic variable derivation."""
